requirements.txt
logging
requests
google-cloud-storage
boto3
google-auth